html/semantics/forms/the-select-element/customizable-select/select-optgroup-animation-restart.html

<!DOCTYPE html>
<title>Test that animations for the select optgroups restart when re-opening select</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<style>
select, select::picker(select) {
  appearance: base-select;
}

optgroup {
  animation: fade-in 1ms forwards;
  opacity: 0;
}

@keyframes fade-in {
  to {
    opacity: 1;
  }
}
</style>

<select>
  <optgroup>
    <option>one</option>
    <option>two</option>
  </optgroup>
</select>

<script>
const select = document.querySelector("select");
const optgroup = document.querySelector("optgroup");

function waitForAnimationEnd(element) {
  return new Promise(resolve => element.addEventListener("animationend", resolve, { once: true }));
}

promise_test(async t => {
  // Open picker and wait for the animation to complete.
  await test_driver.bless();
  assert_equals(getComputedStyle(optgroup).opacity, "0",
    "optgroup opacity should be 0 before opening the picker.");
  select.showPicker();
  assert_true(select.matches(":open"));

  await waitForAnimationEnd(optgroup);
  assert_equals(getComputedStyle(optgroup).opacity, "1",
    "optgroup opacity should be 1 after animation completes on first open.");

  // Close the picker.
  await test_driver.click(select);
  assert_false(select.matches(":open"));

  // Reopen the picker and verify the animation restarts and completes again.
  await test_driver.bless();
  assert_equals(getComputedStyle(optgroup).opacity, "0",
    "optgroup opacity should be 0 before opening the picker a second time.");
  select.showPicker();
  assert_true(select.matches(":open"));

  await waitForAnimationEnd(optgroup);
  assert_equals(getComputedStyle(optgroup).opacity, "1",
    "optgroup opacity should be 1 after animation completes on second open.");

  // Close the picker.
  await test_driver.click(select);
}, "<optgroup> animations should restart when the picker is reopened.");
</script>
